requirements.txt
langgraph>=0.2.0
langchain-google-genai>=1.0.0
google-cloud-bigquery>=3.13.0
pandas>=2.0.0
python-dotenv>=1.0.0 
langchain_core>=0.3.0
db-dtypes==1.2.0

# Additional dependencies for CLI and visualization
click>=8.0.0
rich>=13.0.0
structlog>=23.0.0

# For state management and typing
typing-extensions>=4.0.0